% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSeqFeatureCppSingle.R
\name{countSeqFeatureCppSingle}
\alias{countSeqFeatureCppSingle}
\title{Count kmer from sequences from single direction.}
\usage{
countSeqFeatureCppSingle(RseqDNA, k)
}
\arguments{
\item{RseqDNA}{The DNA sequence.}

\item{k}{The kmer size.}
}
\description{
Count kmer from sequences from single direction using the cpp file.
}
