% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_one_kmer.R
\name{drop_one_kmer}
\alias{drop_one_kmer}
\title{Drop one kmer.}
\usage{
drop_one_kmer(kmerfX_normal, y)
}
\arguments{
\item{kmerfX_normal}{The normalized kmer frequency matrix.}

\item{y}{The label vector, taking values of 0 or 1 for two classes of sequences.}
}
\value{
drop_idx The index of kmer to be dropped.
}
\description{
Select the kmer having the least likelihood of being relevant by WMW test.
}
