% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknockoffX.R
\name{gknockoffX}
\alias{gknockoffX}
\title{Design matrix generation.}
\usage{
gknockoffX(X, Omega)
}
\arguments{
\item{X}{The normalized kmer frequency matrix with one column dropped.}

\item{Omega}{The estimated precision matrix.}
}
\value{
n x (2p) design matrix = [X X_ko]. X_ko is the knockoff matrix.
}
\description{
Function generating the knockoff design matrix.
}
