% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIMI_plus.R
\name{KIMI_plus}
\alias{KIMI_plus}
\title{KIMI for motif identification.}
\usage{
KIMI_plus(X, y, q, alpha = 1)
}
\arguments{
\item{X}{The normalized kmer frequency matrix with one column dropped.}

\item{y}{The label vector, taking values of 0 or 1 for two classes of sequences.}

\item{q}{The target FDR.}

\item{alpha}{The parameter for the weight of L1 penalty.}
}
\value{
The set of indexes of selected kmers.
}
\description{
KIMI for motif identification based on the knockoff plus statistic.
}
