% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_standard_gaussian.R
\name{normalize_standard_gaussian}
\alias{normalize_standard_gaussian}
\title{Normalize into standard gaussian.}
\usage{
normalize_standard_gaussian(kmerfX)
}
\arguments{
\item{kmerfX}{The kmer frequency matrix.}
}
\value{
A normalized kmer frequency matrix. Each column has mean 0 and standard deviation 1.
}
\description{
Normalize each column of the kmer frequency matrix into a vector with mean 0 and standard deviation 1.
}
