% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFasta.R
\name{readFasta}
\alias{readFasta}
\title{Read dasta.}
\usage{
readFasta(FastaDir, w, is_single_strand = TRUE)
}
\arguments{
\item{FastaDir}{The input directory of the fasta/fastq file.}

\item{w}{The kmer size.}

\item{is_single_strand}{A bool variable indicating counting single strand or double strand, default to be TRUE.}
}
\value{
A kmer frequency matrix. Each row corresponds to a contig. Each column corresponds to a kmer.
}
\description{
Reads a fasta/fastq format file containing contigs, count kmer frequency for each contig, and then output a kmer frequency matrix.
}
